Move size util into its own module
